+++ GenHub/Directory.Packages.props
@@ -25,6 +25,7 @@
     <PackageVersion Include="Microsoft.Playwright" Version="1.55.0" />
     <PackageVersion Include="ReactiveUI" Version="19.6.1" />
     <PackageVersion Include="Serilog.Extensions.Logging.File" Version="3.0.0" />
+    <PackageVersion Include="Slugify.Core" Version="5.1.1" />
     <PackageVersion Include="StyleCop.Analyzers" Version="1.1.118" />
     <PackageVersion Include="CsvHelper" Version="33.1.0" />
     <!-- Test packages -->
