+++ GenHub/GenHub/Features/Content/Services/ContentDiscoverers/ModDBDiscoverer.cs
@@ -21,9 +21,6 @@ namespace GenHub.Features.Content.Services.ContentDiscoverers;
 /// </summary>
 public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> logger) : IContentDiscoverer
 {
-    private readonly HttpClient _httpClient = httpClient;
-    private readonly ILogger<ModDBDiscoverer> _logger = logger;
-
     /// <inheritdoc />
     public string SourceName => ModDBConstants.DiscovererSourceName;
 
@@ -44,7 +41,7 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
         try
         {
             var gameType = query.TargetGame ?? GameType.ZeroHour;
-            _logger.LogInformation("Discovering ModDB content for {Game}", gameType);
+            logger.LogInformation("Discovering ModDB content for {Game}", gameType);
 
             var results = new List<ContentSearchResult>();
 
@@ -57,7 +54,7 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
                 results.AddRange(sectionResults);
             }
 
-            _logger.LogInformation(
+            logger.LogInformation(
                 "Discovered {Count} ModDB items across {Sections} sections",
                 results.Count,
                 sectionsToSearch.Count);
@@ -66,7 +63,7 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
         }
         catch (Exception ex)
         {
-            _logger.LogError(ex, "Failed to discover ModDB content");
+            logger.LogError(ex, "Failed to discover ModDB content");
             return OperationResult<IEnumerable<ContentSearchResult>>.CreateFailure($"Discovery failed: {ex.Message}");
         }
     }
@@ -108,17 +105,17 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
         {
             // Build URL for the section
             var baseUrl = gameType == GameType.Generals
-                ? $"{ModDBConstants.BaseUrl}/games/cc-generals/{section}"
-                : $"{ModDBConstants.BaseUrl}/games/cc-generals-zero-hour/{section}";
+                ? $"{ModDBConstants.GeneralsBaseUrl}/{section}"
+                : $"{ModDBConstants.ZeroHourBaseUrl}/{section}";
 
             // Build filter query string if needed
             var filter = BuildFilterFromQuery(query, section);
             var queryString = filter.ToQueryString();
             var url = baseUrl + queryString;
 
-            _logger.LogDebug("Fetching from URL: {Url}", url);
+            logger.LogDebug("Fetching from URL: {Url}", url);
 
-            var html = await _httpClient.GetStringAsync(url, cancellationToken);
+            var html = await httpClient.GetStringAsync(url, cancellationToken);
             var context = BrowsingContext.New(Configuration.Default);
             var document = await context.OpenAsync(req => req.Content(html), cancellationToken);
 
@@ -140,16 +137,16 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
                 }
                 catch (Exception ex)
                 {
-                    _logger.LogDebug(ex, "Failed to parse content item");
+                    logger.LogDebug(ex, "Failed to parse content item");
                 }
             }
 
-            _logger.LogDebug("Found {Count} items in {Section} section", results.Count, section);
+            logger.LogDebug("Found {Count} items in {Section} section", results.Count, section);
             return results;
         }
         catch (Exception ex)
         {
-            _logger.LogWarning(ex, "Failed to discover from {Section} section", section);
+            logger.LogWarning(ex, "Failed to discover from {Section} section", section);
             return [];
         }
     }
@@ -190,11 +187,11 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
         {
             return contentType switch
             {
-                ContentType.Mod => "2", // Full Version
-                ContentType.Patch => "4", // Patch
-                ContentType.Video => "8", // Movie
-                ContentType.ModdingTool => "14", // Mapping Tool
-                ContentType.LanguagePack => "30", // Language Pack
+                ContentType.Mod => ModDBConstants.CategoryFullVersion,
+                ContentType.Patch => ModDBConstants.CategoryPatch,
+                ContentType.Video => ModDBConstants.CategoryMovie,
+                ContentType.ModdingTool => ModDBConstants.CategoryMappingTool,
+                ContentType.LanguagePack => ModDBConstants.CategoryLanguagePack,
                 _ => null
             };
         }
@@ -202,9 +199,9 @@ public class ModDBDiscoverer(HttpClient httpClient, ILogger<ModDBDiscoverer> log
         {
             return contentType switch
             {
-                ContentType.Map => "101", // Multiplayer Map
-                ContentType.Skin => "112", // Player Skin
-                ContentType.LanguagePack => "138", // Language Sounds
+                ContentType.Map => ModDBConstants.AddonMultiplayerMap,
+                ContentType.Skin => ModDBConstants.AddonPlayerSkin,
+                ContentType.LanguagePack => ModDBConstants.AddonLanguageSounds,
                 _ => null
             };
         }
