+++ GenHub/GenHub/GenHub.csproj
@@ -30,6 +30,7 @@
         <PackageReference Include="Octokit" />
         <PackageReference Include="ReactiveUI" />
         <PackageReference Include="Serilog.Extensions.Logging.File" />
+        <PackageReference Include="Slugify.Core" />
         <PackageReference Include="StyleCop.Analyzers" />
         <PackageReference Include="Velopack" />
     </ItemGroup>
